requirements.txt
Django>=3.1.13,<4.0
django-ratelimit==3.0.0
django-extensions==3.0.2
graphviz
pydotplus
sendgrid==6.4.3
sentry-sdk
# just for postgres
psycopg2-binary
django-taggit==1.3.0
djangorestframework==3.11.2
django-rest-swagger
django-crispy-forms==1.11.0
social-auth-core
social-auth-app-django
django-gravatar2
ipython
pyaml
gunicorn
whitenoise
pandas
django-river-ml
creme
django-autocomplete-light
git+https://github.com/vsoch/online-ml-custom.git@main
